requirements.txt
requests
pandas
xgboost
python-telegram-bot==13.15
Flask
joblib
apscheduler